+++ app/src/main/java/de/passwordvault/model/storage/app/StorageManager.java
@@ -85,7 +85,7 @@ public class StorageManager {
             return null;
         }
 
-        //Convert content into CSV instances:
+        //Convert content into DTO instances:
         ArrayList<EntryDTO> entryDTOs = new ArrayList<>();
         ArrayList<DetailDTO> detailDTOs = new ArrayList<>();
         if (entryCsv != null) {