+++ app/src/main/java/de/passwordvault/view/fragments/SettingsFragment.java
@@ -1,9 +1,13 @@
 package de.passwordvault.view.fragments;
 
+import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
@@ -12,8 +16,13 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import android.widget.Toast;
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity;
 import de.passwordvault.R;
+import de.passwordvault.model.storage.backup.BackupException;
+import de.passwordvault.model.storage.backup.CreateXmlBackup;
+import de.passwordvault.model.storage.backup.RestoreXmlBackup;
+import de.passwordvault.model.storage.backup.XmlException;
 import de.passwordvault.view.utils.DialogCallbackListener;
 import de.passwordvault.view.viewmodel.SettingsViewModel;
 import de.passwordvault.view.activities.MainActivity;
@@ -25,9 +34,20 @@ import de.passwordvault.view.dialogs.UiModeDialogFragment;
  * {@linkplain MainActivity}.
  *
  * @author  Christian-2003
- * @version 3.0.0
+ * @version 3.1.0
  */
 public class SettingsFragment extends Fragment implements DialogCallbackListener {
+
+    /**
+     * Field stores the request code to pick a location for the XML file for a backup.
+     */
+    private static final int PICK_XML_FILE_LOCATION = 2;
+
+    /**
+     * Field stores the request coe to pick a location for the XML backup file to be restored.
+     */
+    private static final int PICK_XML_BACKUP_LOCATION = 3;
+
 
     /**
      * Attribute stores the {@linkplain androidx.lifecycle.ViewModel} for this fragment.
@@ -98,8 +118,12 @@ public class SettingsFragment extends Fragment implements DialogCallbackListener
             dialog.setTargetFragment(SettingsFragment.this, 1);
             dialog.show(getActivity().getSupportFragmentManager(), "");
         });
-
+        inflated.findViewById(R.id.settings_xml_data_backup_clickable).setOnClickListener(view -> createXmlBackup());
+        inflated.findViewById(R.id.settings_xml_data_backup_info).setOnClickListener(view -> showInfoDialog(R.string.settings_create_data_backup, R.string.settings_xml_backup_info));
+        inflated.findViewById(R.id.settings_xml_data_restoration_clickable).setOnClickListener(view -> restoreXmlBackup());
+        inflated.findViewById(R.id.settings_xml_data_restoration_info).setOnClickListener(view -> showInfoDialog(R.string.settings_restore_data_backup, R.string.settings_xml_restoration_info));
         inflated.findViewById(R.id.settings_used_software_clickable).setOnClickListener(view -> startActivity(new Intent(getActivity(), OssLicensesMenuActivity.class)));
+        inflated.findViewById(R.id.settings_license_clickable).setOnClickListener(view -> showInfoDialog(R.string.settings_license_notice, R.string.app_license));
 
         return inflated;
     }
@@ -129,6 +153,57 @@ public class SettingsFragment extends Fragment implements DialogCallbackListener
     }
 
 
+    /**
+     * Method is called whenever an activity has finished after being called through
+     * {@link #startActivityForResult(Intent, int)}
+     *
+     * @param requestCode   The integer request code originally supplied to
+     *                      startActivityForResult(), allowing you to identify who this
+     *                      result came from.
+     * @param resultCode    The integer result code returned by the child activity
+     *                      through its setResult().
+     * @param data          An Intent, which can return result data to the caller
+     *                      (various data can be attached to Intent "extras").
+     */
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        if (requestCode == PICK_XML_FILE_LOCATION && resultCode == Activity.RESULT_OK) {
+            //Create backup:
+            if (data != null) {
+                CreateXmlBackup backup = new CreateXmlBackup(data.getData());
+                try {
+                    backup.createBackup();
+                }
+                catch (BackupException e) {
+                    Toast.makeText(getContext(), getString(R.string.settings_backup_error), Toast.LENGTH_LONG).show();
+                    return;
+                }
+                Toast.makeText(getContext(), getString(R.string.settings_backup_success), Toast.LENGTH_SHORT).show();
+            }
+        }
+        else if (requestCode == PICK_XML_BACKUP_LOCATION && resultCode == Activity.RESULT_OK) {
+            //Restore backup:
+            if (data != null) {
+                RestoreXmlBackup backup = new RestoreXmlBackup(data.getData());
+                try {
+                    backup.restoreBackup();
+                }
+                catch (BackupException | XmlException e) {
+                    e.printStackTrace();
+                    Toast.makeText(getContext(), e.getMessage(), Toast.LENGTH_LONG).show();
+                    return;
+                }
+                Toast.makeText(getContext(), getString(R.string.settings_restoration_success), Toast.LENGTH_SHORT).show();
+            }
+        }
+    }
+
+
+    /**
+     * Method updates the UI mode of the application.
+     *
+     * @param uiMode    UI mode to which the app shall be changed.
+     */
     private void updateUiMode(int uiMode) {
         if (viewModel.getUiMode() == uiMode) {
             //UI mode was not changed:
@@ -156,6 +231,43 @@ public class SettingsFragment extends Fragment implements DialogCallbackListener
                 Log.d(TAG, "Changed UI mode to: SYSTEM DEFAULT MODE");
                 break;
         }
+    }
+
+
+    /**
+     * Method creates an XML backup for the application.
+     */
+    private void createXmlBackup() {
+        Intent intent = new Intent(Intent.ACTION_CREATE_DOCUMENT);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        intent.setType("application/*");
+        intent.putExtra(Intent.EXTRA_TITLE, getString(R.string.settings_default_backup_name));
+        startActivityForResult(intent, PICK_XML_FILE_LOCATION);
+    }
+
+    /**
+     * Method restores an XML backup for the application.
+     */
+    private void restoreXmlBackup() {
+        Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        intent.setType("application/*");
+        startActivityForResult(intent, PICK_XML_BACKUP_LOCATION);
+    }
+
+
+    /**
+     * Method displays an information dialog.
+     *
+     * @param titleId   Id of the resource-string for the dialog title.
+     * @param messageId Id if the resource-string for the dialog message.
+     */
+    private void showInfoDialog(int titleId, int messageId) {
+        AlertDialog dialog = new AlertDialog.Builder(getActivity()).create();
+        dialog.setTitle(getString(titleId));
+        dialog.setMessage(getString(messageId));
+        dialog.setButton(DialogInterface.BUTTON_NEUTRAL, getString(R.string.button_ok), (dialogInterface, i) -> dialogInterface.dismiss());
+        dialog.show();
     }
 
 }